implemented redirecting to login page when no user is logged in
