+++ src/App.jsx
@@ -1,11 +1,11 @@
 import React, { createContext, useState } from "react";
 import { Routes, Route } from "react-router-dom";
 import "./App.css";
-import Navbar from "./Components/Navbar";
-import Home from "./Pages/Home";
-import Order from "./Pages/Order";
-import Profile from "./Pages/Profile";
-import LoginRegister from "./Components/LoginRegister/LoginRegister";
+import Navbar from "./components/Navbar";
+import Home from "./pages/Home";
+import Order from "./pages/Order";
+import Profile from "./pages/Profile";
+import LoginRegister from "./components/LoginRegister/LoginRegister";
 
 export const UserContext = createContext();
