--- src/Pages/Home.jsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-function Home() {
-  return (
-    <div>
-      <h1>This is the Home Page</h1>
-    </div>
-  )
-}
-
-export default Home;
