--- src/Pages/Order.jsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-function Order() {
-  return (
-    <div>
-      <h1>This is the Order Page</h1>
-    </div>
-  )
-}
-
-export default Order
