--- src/Pages/Profile.jsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-function Profile() {
-  return (
-    <div>
-      <h1>This is the Profile Page</h1>
-    </div>
-  )
-}
-
-export default Profile
