+++ src/pages/Profile.jsx
@@ -1,0 +1,22 @@
+import React, { useEffect, useContext } from "react";
+import { useNavigate } from "react-router-dom";
+import { UserContext } from "../App";
+
+function Profile() {
+  const navigate = useNavigate();
+  const [userId, setUserId] = useContext(UserContext);
+
+  useEffect(() => {
+    if (!userId) {
+      navigate("/login");
+    }
+  }, []);
+
+  return (
+    <div>
+      <h1>This is the Profile Page</h1>
+    </div>
+  );
+}
+
+export default Profile;
